+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 This is used for recording Invader's changes. This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- invader-bitmap: Spacing now attempts to sort both vertical and horizontal to see if sprites will fit in a sprite sheet
+
 ### Fixed
 - invader-bitmap: Fixed some issues with spacing